refactor(node): Has* accessor traits and incentive API (#44)
## Summary

- Replace monolithic HoprNodeChainOperations and
HoprNodeTicketOperations with composable Has* accessor traits
- Each accessor exposes a typed component reference and sync status()
for per-component health
- Blanket-impl IncentiveChannelOperations (channels + balances, edge
nodes) and IncentiveRedeemOperations (ticket redemption, relay nodes)
- Add TransportOperations trait (ping, observed multiaddresses)
- Add HoprGraphReadApi (read-only graph composite)
- Remove PeerId from node module public API
- Remove HoprNodeNetworkOperations (consumers use NetworkView directly)
- Remove noop `impl TicketManagement for ()` — edge nodes use `TMgr=()`
structurally but never invoke TicketManagement methods (enabled by
hoprnet PR #8055)
- Add `auto_impl(&, Arc)` to all Has* accessor traits including
HasTicketManagement
- Fix unused import in node/types.rs
- Version bump to 1.8.0

Depends on: hoprnet/hoprnet#8053 (companion PR with Has* impls on Hopr
struct)

---------
