+++ src/ct/mod.rs
@@ -1,3 +1,8 @@
+//! Cover traffic and probing traffic generation traits.
+//!
+//! - [`ProbingTrafficGeneration`] — network probe scheduling and execution
+//! - [`CoverTrafficGeneration`] — cover traffic generation for privacy
+
 pub mod traits;
 
 pub use traits::*;
