+++ src/lib.rs
@@ -1,3 +1,24 @@
+//! Common high-level external and internal API traits for the HOPR protocol.
+//!
+//! This crate defines **trait-based interfaces** that separate API contract from implementation.
+//! Concrete implementations live in their respective crates (`hopr-lib`, `hopr-transport`, etc.)
+//! and depend on the traits defined here.
+//!
+//! ## Module Organization
+//!
+//! - [`chain`] — On-chain operations: channel management, account queries, safe operations, event subscriptions
+//! - [`ct`] — Cover traffic and probing traffic generation
+//! - [`graph`] — Network graph: topology view, pathfinding, edge quality observations
+//! - [`network`] — Network layer: peer connectivity, health, stream control
+//! - [`node`] — High-level node API: accessor traits (`Has*`), composed operations, session client
+//! - [`tickets`] — Winning ticket management and redemption
+//!
+//! ## Design Principle
+//!
+//! The interface mandates trait behavior defined in this crate and does not rely on
+//! specific types outside of this crate. External types (from `hopr-types`) are
+//! re-exported at the crate root for convenience.
+
 /// On-chain operations-related API traits.
 #[cfg(feature = "chain")]
 pub mod chain;
