+++ src/network/traits.rs
@@ -1,15 +1,15 @@
 use std::collections::HashSet;
 
 use futures::{AsyncRead, AsyncWrite, Stream, future::BoxFuture};
-use hopr_types::crypto::keypairs::OffchainKeypair;
 
-use super::Health;
+use super::{Health, NetworkEvent};
 use crate::{Multiaddr, PeerId};
 
 /// Type alias for a boxed function returning a boxed future.
 pub type BoxedProcessFn = Box<dyn FnOnce() -> BoxFuture<'static, ()> + Send>;
 
 /// Trait representing a read-only view of the network state.
+#[auto_impl::auto_impl(&, Arc)]
 pub trait NetworkView {
     /// Multiaddresses used for listening by the local node.
     fn listening_as(&self) -> HashSet<Multiaddr>;
@@ -28,6 +28,12 @@ pub trait NetworkView {
 
     /// Represents perceived health of the network.
     fn health(&self) -> Health;
+
+    /// Subscribes to network events (peer connected/disconnected).
+    ///
+    /// Each call creates a new independent subscription. Must be called
+    /// before the network starts processing to avoid missing initial events.
+    fn subscribe_network_events(&self) -> impl Stream<Item = NetworkEvent> + Send + 'static;
 }
 
 /// Control object for the opening and receiving of network connections in the
@@ -39,21 +45,4 @@ pub trait NetworkStreamControl: std::fmt::Debug {
     ) -> Result<impl Stream<Item = (PeerId, impl AsyncRead + AsyncWrite + Send)> + Send, impl std::error::Error>;
 
     async fn open(self, peer: PeerId) -> Result<impl AsyncRead + AsyncWrite + Send, impl std::error::Error>;
-}
-
-/// Builder for a network object that returns the network specified in
-/// the `NetworkBuilder::Network` associated type.
-#[async_trait::async_trait]
-pub trait NetworkBuilder {
-    type Network: NetworkView + NetworkStreamControl + Send + Sync + Clone + 'static;
-
-    /// Build and return an instance of a network viewer along with a process to start
-    /// and run the network component.
-    async fn build(
-        self,
-        identity: &OffchainKeypair,
-        my_multiaddresses: Vec<Multiaddr>,
-        protocol: &'static str,
-        allow_private_addresses: bool,
-    ) -> Result<(Self::Network, BoxedProcessFn), impl std::error::Error>;
 }
